+++ src/main/java/antigravity/controller/api/ProductController.java
@@ -1,4 +1,4 @@
-package antigravity.controller;
+package antigravity.controller.api;
 
 public class ProductController {
